--- datasets/dna.py
@@ -1,12 +1,0 @@
-from torch.utils.data import Dataset
-
-class DnaDataset(Dataset):
-    def __init__(self, data, target):
-        self.data = data
-        self.target = target
-
-    def __len__(self):
-        return self.x.shape[0]
-
-    def __getitem__(self, idx):
-        return self.x[idx], self.y[idx]
