+++ params.ini
@@ -1,10 +1,15 @@
 [DEAFULT]
 
+
 [DNA]
-train = data\dna.tr
-val = data\dna.val
-test = data\dna.t
+train = dna/dna_train_raw
+val = dna/dna_val_raw
+test = dna/dna_test_raw
+train_url="https://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/multiclass/dna.scale.tr"
+val_url="https://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/multiclass/dna.scale.val"
+test_url="https://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/multiclass/dna.scale.t"
 n_features = 180
+n_claeses = 3
 
 [SPLICE]
 train = data\splice.tr
